+++ LinkedIn/src/database/dao/job/JobDAO.java
@@ -8,15 +8,15 @@ public interface JobDAO {
 	
 	public List<Job> list();
 
-    public int create(Job job, Long id);
+    public int create(Job job, int id);
     
     public int count();
     
-    public Job findJob(Long id);
+    public Job findJob(int id);
     
-    public int updateJob(Job job, Long id);
+    public int updateJob(Job job, int id);
     
-    public List<Job> getSessionJobs(Long userId);
+    public List<Job> getSessionJobs(int userId);
     
-    public List<Job> getConnectionsJobs(Long userId);
+    public List<Job> getConnectionsJobs(int userId);
 }
